+++ src/win.h
@@ -16,6 +16,7 @@ void win_bell(void);
 void win_set_title(char *);
 void win_save_title(void);
 void win_restore_title(void);
+void win_prefix_title(const char *);
 
 colour win_get_colour(colour_i);
 void win_set_colour(colour_i, colour);