+++ wiki/Changelog.md
@@ -1,3 +1,4 @@
+  * Added termination indicator config options ExitWrite and ExitTitle (#437).
   * Added options -B frame and -B void to reduce window border (#7).
   * Displaying child process list on exit confirmation (#448).